agent_engine_agents/iam_classic_joke_agent/pyproject.toml
[project]
name = "classic_joke_agent_iam"
version = "0.1.0"
dependencies = [
    "google-adk",
    "requests",
    "google-auth"
]